Formatting?

+++ src/main/java/net/ilexiconn/llibrary/client/gui/survivaltab/PageButtonGUI.java
@@ -24,7 +24,7 @@ public class PageButtonGUI extends GuiButton {
     public void drawButton(Minecraft mc, int mouseX, int mouseY) {
         super.drawButton(mc, mouseX, mouseY);
         if (this.id == -1) {
-            mc.fontRendererObj.drawString((SurvivalTabHandler.INSTANCE.getCurrentPage() + 1) + "/" + (int) (SurvivalTabHandler.INSTANCE.getSurvivalTabList().size() / 8 + 1), this.xPosition + 31, this.yPosition + 6, 0xFFFFFFFF);
+            mc.fontRendererObj.drawString((SurvivalTabHandler.INSTANCE.getCurrentPage() + 1) + "/" + SurvivalTabHandler.INSTANCE.getSurvivalTabList().size() / 8 + 1, this.xPosition + 31, this.yPosition + 6, 0xFFFFFFFF);
         }
     }
